Fix the error that caused the renderer to crash when loading a new scene
The error occurred because all the materials seem to be destroyed when loading a new scene despite the fact that someting can still be referencing them. The render feature was instantiating it's rendering material in the Create() method which is only called once when loading the game so when the rendering material got destroyed when loading a different scene thre was nothing that could create a new one.

The error was resolved by moving the material creation out of the InscatterinfPass constructor to the AddRenderPasses() method of the Inscattering render feature